2021/12/31

+++ src/main/java/com/huanhong/wms/controller/WarehouseAreaManagementController.java
@@ -9,15 +9,20 @@ import com.baomidou.mybatisplus.core.toolkit.StringUtils;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.github.xiaoymin.knife4j.annotations.ApiOperationSupport;
 import com.github.xiaoymin.knife4j.annotations.ApiSort;
+import com.huanhong.common.units.StrUtils;
 import com.huanhong.wms.BaseController;
 import com.huanhong.wms.bean.ErrorCode;
 import com.huanhong.wms.bean.Result;
 import com.huanhong.wms.config.JudgeConfig;
+import com.huanhong.wms.entity.ShelfManagement;
+import com.huanhong.wms.entity.SublibraryManagement;
 import com.huanhong.wms.entity.WarehouseAreaManagement;
 import com.huanhong.wms.entity.dto.AddWarehouseAreaDTO;
 import com.huanhong.wms.entity.dto.UpdateWarehouseAreaDTO;
 import com.huanhong.wms.entity.vo.WarehouseAreaVO;
 import com.huanhong.wms.mapper.WarehouseAreaManagementMapper;
+import com.huanhong.wms.service.IShelfManagementService;
+import com.huanhong.wms.service.ISublibraryManagementService;
 import com.huanhong.wms.service.IWarehouseAreaManagementService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiImplicitParam;
@@ -43,6 +48,12 @@ public class WarehouseAreaManagementController extends BaseController {
 
     @Resource
     private WarehouseAreaManagementMapper warehouseAreaManagementMapper;
+
+    @Resource
+    private ISublibraryManagementService sublibraryManagementService;
+
+    @Resource
+    private IShelfManagementService shelfManagementServicel;
 
     @Autowired
     private JudgeConfig judgeConfig;
@@ -70,14 +81,35 @@ public class WarehouseAreaManagementController extends BaseController {
     }
 
     @ApiOperationSupport(order = 2)
-    @ApiOperation(value = "添加库房区域管理", notes = "生成代码")
-    @PostMapping
+    @ApiOperation(value = "添加库区")
+    @PostMapping("/add")
     public Result add(@Valid @RequestBody AddWarehouseAreaDTO addWarehouseAreaDTO) {
 
         /**
          * 判断是否有必填参数为空
          */
         try {
+
+            /**
+             * 查询子库是否存在
+             */
+            SublibraryManagement sublibraryManagement = sublibraryManagementService.getSublibraryBySublibraryId(addWarehouseAreaDTO.getSublibraryId());
+            if (ObjectUtil.isEmpty(sublibraryManagement)) {
+                return Result.failure(ErrorCode.DATA_IS_NULL, "子库不存在");
+            }
+
+            /**
+             *验证库区编码
+             */
+
+            if(addWarehouseAreaDTO.getWarehouseAreaId().length()!=1||!StrUtils.isEnglish(addWarehouseAreaDTO.getWarehouseAreaId())){
+                return Result.failure(ErrorCode.DATA_IS_NULL, "库区编号为除I和O以外的一位大写字母");
+            }
+
+            /**
+             * 组合子库编号和库区编号为完整库区编号
+             */
+            addWarehouseAreaDTO.setWarehouseAreaId(addWarehouseAreaDTO.getSublibraryId()+addWarehouseAreaDTO.getWarehouseAreaId());
             /**
              * 实体类转为json
              */
@@ -131,8 +163,8 @@ public class WarehouseAreaManagementController extends BaseController {
     }
 
     @ApiOperationSupport(order = 3)
-    @ApiOperation(value = "更新库房区域管理", notes = "生成代码")
-    @PutMapping
+    @ApiOperation(value = "更新库区")
+    @PutMapping("/update")
     public Result update(@Valid @RequestBody UpdateWarehouseAreaDTO updateWarehouseAreaDTO) {
         UpdateWrapper updateWrapper = new UpdateWrapper<>();
         try {
@@ -157,8 +189,8 @@ public class WarehouseAreaManagementController extends BaseController {
     }
 
     @ApiOperationSupport(order = 4)
-    @ApiOperation(value = "删除库房区域管理", notes = "生成代码")
-    @DeleteMapping("/{warehouseAreaId}")
+    @ApiOperation(value = "删除库区")
+    @DeleteMapping("delete/{warehouseAreaId}")
     public Result delete(@PathVariable String warehouseAreaId) {
         QueryWrapper<WarehouseAreaManagement> queryWrapper = new QueryWrapper<>();
         try {
@@ -167,12 +199,44 @@ public class WarehouseAreaManagementController extends BaseController {
             if (ObjectUtil.isEmpty(warehouseAreaManagement)) {
                 return Result.failure(ErrorCode.SYSTEM_ERROR, "操作失败：库区编号不存在");
             }
+            /**
+             * 检查此库区下是否有货架
+             */
+            List<ShelfManagement> shelfManagementList =  shelfManagementServicel.getShelfListByWarehouseAreaId(warehouseAreaId);
+            if (ObjectUtil.isNotEmpty(shelfManagementList)){
+                return Result.failure(ErrorCode.SYSTEM_ERROR, "此库区下存在货架，请先删除货架");
+            }
+
             int i = warehouseAreaManagementMapper.delete(queryWrapper);
             LOGGER.info("库区:  " + warehouseAreaId + "删除成功");
             return render(i > 0);
         } catch (Exception e) {
             LOGGER.error("删除库区信息出错--删除失败，异常：" + e);
             return Result.failure(ErrorCode.SYSTEM_ERROR, "系统异常：库区删除失败，请稍后再试或联系管理员");
+        }
+    }
+
+    @ApiOperationSupport(order = 5)
+    @ApiOperation(value = "获取库区详细信息")
+    @GetMapping("/getWarehouseAreaByWarehouseAreaId/{warehouseAreaId}")
+    public Result getWarehouseArea(@PathVariable String warehouseAreaId) {
+        try {
+            WarehouseAreaManagement warehouseAreaManagement = warehouseAreaManagementService.getWarehouseAreaByWarehouseAreaId(warehouseAreaId);
+            return Result.success(warehouseAreaManagement);
+        } catch (Exception e) {
+            return Result.failure(ErrorCode.SYSTEM_ERROR, "获取库区信息失败");
+        }
+    }
+
+    @ApiOperationSupport(order = 6)
+    @ApiOperation(value = "获取对应库区的所有货架")
+    @GetMapping("/getShelfByWarehouseId/{warehouseAreaId}")
+    public Result getAllShelf(@PathVariable String warehouseAreaId) {
+        try {
+            List<ShelfManagement> ShelfManagementList = shelfManagementServicel.getShelfListByWarehouseAreaId(warehouseAreaId);
+            return Result.success(ShelfManagementList);
+        } catch (Exception e) {
+            return Result.failure(ErrorCode.SYSTEM_ERROR, "获取货架信息失败");
         }
     }
 }
